requirements.txt
flwr

torch

torchvision

numpy

transformers>=4.37.0

peft

accelerate

fastapi

uvicorn

streamlit

datasets

distilabel

python-dotenv